git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@14159 01821946-851b-0410-b7f4-f1477c70e57a

+++ effects/lightning_manager.cpp
@@ -269,7 +269,7 @@ void CLightningManager::Update (void)
 if (SHOW_LIGHTNING (1)) {
 
 		CObject	*pObj;
-		uint8_t		h;
+		uint16_t		h;
 		int32_t		i;
 
 #if LIMIT_LIGHTNING_FPS
